# docs/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: FreshLens API
  version: 1.0.0
  description: |
    V1 contract for FreshLens (CS3203 · Group 21 · PID 5).

    - Base path for versioned resources: `/api/v1`
    - Auth: Supabase JWT Bearer on versioned routes except public tenant
      application submission; `GET /health` is also public
    - Tenant isolation: auth middleware derives trusted identity from the JWT, and
      the tenant DB transaction sets Postgres `app.tenant_id` on its own connection;
      never trust a client-supplied `tenant_id`
    - Async scan only: `POST /api/v1/scans` returns **202**; the two-tier CNN
      pipeline runs in Celery, never in the request handler

    See `docs/api/README.md` for human overview and out-of-scope notes.
servers:
  - url: http://localhost:8000
    description: Local Docker Compose / FastAPI (prototype)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tenant/overview:
    get:
      tags: [Tenant workspace]
      summary: Get the current tenant's operational overview
      operationId: getTenantOverview
      x-freshlens-roles: [tenant_admin]
      responses:
        "200":
          description: RLS-scoped tenant totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/tenant/analytics:
    get:
      tags: [Tenant workspace]
      summary: Get aggregate scan analytics for the current tenant
      operationId: getTenantAnalytics
      x-freshlens-roles: [tenant_admin]
      parameters:
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 90, default: 30 }
      responses:
        "200":
          description: Tenant-scoped daily and pipeline totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnalytics"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/tenant/users:
    get:
      tags: [Tenant workspace]
      summary: List vendor users in the current tenant
      operationId: listCurrentTenantUsers
      x-freshlens-roles: [tenant_admin]
      responses:
        "200":
          description: Current tenant's vendor users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Tenant workspace]
      summary: Invite a vendor into the current tenant
      operationId: createCurrentTenantUser
      x-freshlens-roles: [tenant_admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUserCreate"
      responses:
        "201":
          description: Vendor account mapped and invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Email is already registered
  /api/v1/tenant/users/{user_id}/status:
    patch:
      tags: [Tenant workspace]
      summary: Allow or revoke one vendor in the current tenant
      operationId: updateCurrentTenantUserStatus
      x-freshlens-roles: [tenant_admin]
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessStatusUpdate"
      responses:
        "200":
          description: Updated vendor access status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/admin/tenants/{tenant_id}/users:
    get:
      tags: [Admin]
      summary: List vendor users in a tenant
      operationId: listTenantUsers
      security:
        - bearerAuth: []
      x-freshlens-roles: [platform_admin]
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
      responses:
        "200":
          description: Tenant user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Tenant not found
    post:
      tags: [Admin]
      summary: Invite a new vendor user into an existing tenant
      operationId: createTenantUser
      security:
        - bearerAuth: []
      x-freshlens-roles: [platform_admin]
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUserCreate"
      responses:
        "201":
          description: User mapped to the tenant and invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Tenant not found
        "409":
          description: Email is already registered
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/admin/tenants/{tenant_id}/users/{user_id}/status:
    patch:
      tags: [Admin]
      summary: Allow or revoke one tenant user's access
      operationId: updateTenantUserStatus
      security:
        - bearerAuth: []
      x-freshlens-roles: [platform_admin]
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
        - name: user_id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessStatusUpdate"
      responses:
        "200":
          description: Updated user access status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Tenant user not found
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token (JWT).
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but wrong role or tenant
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found in this tenant
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ScanStatus:
      type: string
      enum: [pending, processing, completed, failed]
...
    Error:
      type: object
      required: [detail]
      additionalProperties: false
      properties:
        detail:
          description: FastAPI-compatible error payload (string or structured object).
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...
    TenantUserCreate:
      type: object
      required: [display_name, email]
      additionalProperties: false
      properties:
        display_name: { type: string, minLength: 1, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
    TenantUser:
      type: object
      required: [id, tenant_id, display_name, email, status, created_at, updated_at]
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        tenant_id:
          $ref: "#/components/schemas/Uuid"
        display_name: { type: string }
        email: { type: string, format: email }
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TenantUserCreated:
      type: object
      required: [id, tenant_id, display_name, email, status, created_at, updated_at, invitation_sent]
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        tenant_id:
          $ref: "#/components/schemas/Uuid"
        display_name: { type: string }
        email: { type: string, format: email }
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        invitation_sent: { type: boolean, const: true }
    TenantUserList:
      type: object
      required: [items, total]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TenantUser"
        total:
          type: integer
          minimum: 0
    AccessStatusUpdate:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [active, inactive]
...
    TenantOverview:
      type: object
      additionalProperties: false
      required: [tenant_name, tenant_status, team_members, catalogue_products, active_batches, units_in_stock, active_alerts, scans_this_month, fresh_scans_this_month, medium_scans_this_month, spoiled_scans_this_month]
      properties:
        tenant_name: { type: string }
        tenant_status: { type: string, enum: [active, inactive] }
        team_members: { type: integer, minimum: 0 }
        catalogue_products: { type: integer, minimum: 0 }
        active_batches: { type: integer, minimum: 0 }
        units_in_stock: { type: integer, minimum: 0 }
        active_alerts: { type: integer, minimum: 0 }
        scans_this_month: { type: integer, minimum: 0 }
        fresh_scans_this_month: { type: integer, minimum: 0 }
        medium_scans_this_month: { type: integer, minimum: 0 }
        spoiled_scans_this_month: { type: integer, minimum: 0 }
...
    AdminTrendPoint:
      type: object
      required: [date, scans, fresh, medium, spoiled]
      additionalProperties: false
      properties:
        date:
          type: string
          format: date
        scans:
          type: integer
          minimum: 0
        fresh:
          type: integer
          minimum: 0
        medium:
          type: integer
          minimum: 0
        spoiled:
          type: integer
          minimum: 0
    AdminPipelineTotal:
      type: object
      required: [status, count]
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/ScanStatus"
        count:
          type: integer
          minimum: 0
    AdminAnalytics:
      type: object
      required: [days, tenant_id, trend, pipeline]
      additionalProperties: false
      properties:
        days:
          type: integer
          minimum: 1
          maximum: 90
        tenant_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        trend:
          type: array
          items:
            $ref: "#/components/schemas/AdminTrendPoint"
        pipeline:
          type: array
          items:
            $ref: "#/components/schemas/AdminPipelineTotal"
...